current and voltage readouts working correctly,

+++ src/pm_struct.h
@@ -20,9 +20,9 @@ struct ADC_DATA {
   int32_t adcRaw   = 0;                 // raw value
   int32_t adcMin   = 0;                 // raw value
   int32_t adcMax   = 0;                 // raw value
-  float   adcFloat = 0.0;               // formatted value
-  float   minFloat = 0.0;               // formatted value
-  float   maxFloat = 0.0;               // formatted value
+  double  adcFloat = 0.0;               // formatted value
+  double  minFloat = 0.0;               // formatted value
+  double  maxFloat = 0.0;               // formatted value
 };
 
 volatile I2C_RX_DATA rxData;
